fix: move command_line to separate crate
Closes: #303

Currently, by including `command_line.rs` in `build.rs`, the former gets
snatched into crate "build_script_build".

This has a number of funny and unintuitive side effects like
dependencies not working for the file but the build dependencies.

Here we move `command_line.rs` into a separate "cli" crate. Technically,
this might not be much of an improvement as we still have to add
dependencies to the new crate, but it makes the behaviour more obvious.